ITAM-Server/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
Pillow
python-jose[cryptography]
pandas
openpyxl
reportlab
httpx
pysnmp-lextudio
apscheduler